+++ lib/challenge.js
